pyproject.toml
[project]
name = "pyobs-zwoeaf"
version = "1.0.0"
description = "pyobs model for ZWO EAF focus motors"
authors = [
    { name = "Lennart Jonas Buhrke", email = "l.buhrke@stud.uni-goettingen.de" },
    { name = "Tim-Oliver Husser", email = "thusser@uni-goettingen.de" },
]
requires-python = ">=3.11,<3.14"
license = "MIT"
dependencies = [
    "astropy>=7.0.1,<8",
    "numpy>=2.2.5,<3",
    "pyobs-core>=1,<2",
]

[dependency-groups]
dev = [
    "black>=23.7.0,<24",
    "pre-commit>=3.3.3,<4",
    "flake8>=7.3.0",
    "mypy>=1.17.1",
]

[project.scripts]
pyobs-zwoeaf = "pyobs_zwoeaf.cli:main"

[build-system]
requires = ["scikit-build-core>=0.11", "pybind11>=3.0"]
build-backend = "scikit_build_core.build"

[tool.black]
line-length = 120
target-version = ['py311']
